Update caching.lua
An attempt to fix a problem in the WotLK version after update 3.4.2 - when the bank frame was automatically closed the cache of the character bank items was overwritten with empty data.

Added verification logic so that the bank data cache is saved only on the first triggering of the BANKFRAME_CLOSED event (details in the comments).

Functionality tested in WotLK version 3.4.2.50664 and DF version 10.1.5.50747, but further testing required.